Added and moved color picker for Cylinders and planes
