+++ src/minirt.c
@@ -108,12 +108,12 @@ void	test_scene(void)
 		);
 
 	create_light(
-		make_point3(-7.0, 10.0, 15.0),
-		0.9,
+		make_point3(0.0, 1000.0, 10.0),
+		0.6,
 		make_color(1.0, 1.0, 1.0)
 	);
 
-	create_ambient(0.2, make_color(1.0, 1.0, 1.0));
+	create_ambient(0.1, make_color(1.0, 1.0, 1.0));
 
 	create_camera(
 		make_point3(0.0, 0.0, 0.0),
